change the type of the value within @Order to double

+++ aspectj5rt/java5-src/org/aspectj/lang/annotation/Order.java
@@ -1,6 +1,5 @@
 package org.aspectj.lang.annotation;
 
 public @interface Order {
-
-	int value();
+	double value();
 }
